Add patch from Fedora to skip final cleanup during exit
This has been implicated in various crashes during exit, which cause
gnome-shell to disable extensions during the next startup. Leaking some
memory at this point does not matter since the process is exiting anyway.

The crashes during exit are believed to have been fixed in 44.beta, but
those changes are too intrusive to be suitable for a backport.

Closes: #1038972